feat: add static params
